Rename hooks folder to utils
