+++ webclient/src/components/admin/categories/Banner.js
@@ -3,7 +3,7 @@ import { useMutation, queryCache } from 'react-query';
 import M from 'materialize-css';
 import axios from 'axios';
 import FileUpload from '../utils/FileUpload';
-import { client } from '~/hooks/client';
+import { client } from '~/utils/client';
 import { QUERY_KEY_ADMIN_CATEGORY_DETAILS } from '~/constants/query-keys';
 import { API_ADMIN_CATEGORY, API_FILE_UPLOAD } from '~/constants/api-url';
