+++ webclient/src/pages/Category.js
@@ -6,7 +6,7 @@ import CategoryPeople from '../components/category/CategoryPeople';
 import Carousel from '../components/Carousel';
 import NotFound from './404';
 import Loading from '~/components/Loading';
-import { client } from '~/hooks/client';
+import { client } from '~/utils/client';
 import { QUERY_KEY_CATEGORY } from '~/constants/query-keys';
 import { API_CATEGORIES } from '~/constants/api-url';
