+++ webclient/webpack.common.js
@@ -29,10 +29,6 @@ module.exports = {
   resolve: {
     alias: {
       "~" : path.resolve(__dirname, 'src/'),
-      // constants: path.resolve(__dirname, 'src/constants/'),
-      // css: path.resolve(__dirname, 'src/css/'),
-      // hooks: path.resolve(__dirname, 'src/hooks/'),
-      // pages: path.resolve(__dirname, 'src/pages/'),
     }
   },
   module: {
